+++ src/hooks/useDataUrl.ts
@@ -1,18 +1,24 @@
 import path from 'path';
 
-import { CollectionBeforeChangeHook } from 'payload/types';
+import { CollectionAfterChangeHook } from 'payload/types';
 import sharp from 'sharp';
 
-const useDataUrl: CollectionBeforeChangeHook = async ({ operation, data }) => {
-  if ((operation === 'create' || operation === 'update') && data.mimeType.includes('image')) {
-    const filePath = path.resolve(__dirname, `../../media/${data.filename}`);
+const useDataUrl: CollectionAfterChangeHook = async ({ req: { payload }, operation, doc }) => {
+  if (operation === 'create' || operation === 'update') {
+    const filePath = path.resolve(__dirname, `../../media/${doc.filename}`);
     const imageBuffer = await sharp(filePath).resize(50).toBuffer();
-    const dataUrl = `data:${data.mimeType};base64,${imageBuffer.toString('base64')}`;
+    const dataUrl = `data:${doc.mimeType};base64,${imageBuffer.toString('base64')}`;
 
-    return { ...data, dataUrl };
+    payload.update({
+      collection: 'media',
+      id: doc.id,
+      data: {
+        dataUrl,
+      },
+    });
   }
 
-  return data;
+  return doc;
 };
 
 export default useDataUrl;